1. Adding an ADKMultiGradientView to present a complicated gradient view effect. 2. Refining ADKMultiGradientView and add redraw funtion into it. 3. Adding a example to show how to use ADKGradientView and ADKMultiGradientView. Fixing few samples for layout issues also.
