+++ AppDevKit/GradientViewController.h
@@ -1,0 +1,17 @@
+//
+//  GradientViewController.h
+//  AppDevKit
+//
+//  Created by  Chih Feng Sung on 1/29/19.
+//  Copyright © 2019 Yahoo. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface GradientViewController : UIViewController
+
+@end
+
+NS_ASSUME_NONNULL_END
