+++ AppDevKit/ViewAutoLayoutSupportViewController.xib
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="12121" systemVersion="16F73" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
     <device id="retina4_7" orientation="portrait">
         <adaptation id="fullscreen"/>
     </device>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="12089"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
@@ -23,7 +23,7 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="KqH-WB-0bL">
-                    <rect key="frame" x="0.0" y="223" width="375" height="444"/>
+                    <rect key="frame" x="0.0" y="167" width="375" height="500"/>
                     <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="uQH-XI-uDQ">
                         <size key="itemSize" width="50" height="50"/>
                         <size key="headerReferenceSize" width="0.0" height="0.0"/>
@@ -32,7 +32,7 @@
                     </collectionViewFlowLayout>
                 </collectionView>
                 <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="bxY-CQ-iu9">
-                    <rect key="frame" x="0.0" y="115" width="375" height="100"/>
+                    <rect key="frame" x="0.0" y="59" width="375" height="100"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="100" id="dDu-aw-EIa"/>
                     </constraints>
@@ -44,7 +44,7 @@
                     </collectionViewFlowLayout>
                 </collectionView>
                 <segmentedControl opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="top" segmentControlStyle="plain" selectedSegmentIndex="0" translatesAutoresizingMaskIntoConstraints="NO" id="Hp2-US-kQQ">
-                    <rect key="frame" x="71" y="76" width="233" height="29"/>
+                    <rect key="frame" x="71" y="20" width="233" height="29"/>
                     <segments>
                         <segment title="Complicated List"/>
                         <segment title="Vertical List"/>
@@ -59,14 +59,13 @@
                 <constraint firstAttribute="trailing" secondItem="KqH-WB-0bL" secondAttribute="trailing" id="1HW-Nf-rjr"/>
                 <constraint firstItem="bxY-CQ-iu9" firstAttribute="top" secondItem="Hp2-US-kQQ" secondAttribute="bottom" constant="11" id="4xL-HU-ESn"/>
                 <constraint firstItem="bxY-CQ-iu9" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="5dL-Dk-IxU"/>
-                <constraint firstItem="Hp2-US-kQQ" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" constant="76" id="N4g-i3-Pv4"/>
+                <constraint firstItem="Hp2-US-kQQ" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" constant="20" id="N4g-i3-Pv4"/>
                 <constraint firstAttribute="centerX" secondItem="Hp2-US-kQQ" secondAttribute="centerX" id="Vvw-dc-eR8"/>
                 <constraint firstAttribute="trailing" secondItem="bxY-CQ-iu9" secondAttribute="trailing" id="lBs-8q-EQ9"/>
                 <constraint firstItem="KqH-WB-0bL" firstAttribute="top" secondItem="bxY-CQ-iu9" secondAttribute="bottom" constant="8" id="n9Y-5P-vAs"/>
                 <constraint firstAttribute="bottom" secondItem="KqH-WB-0bL" secondAttribute="bottom" id="uLZ-ua-LLj"/>
                 <constraint firstItem="KqH-WB-0bL" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="zue-Vg-EDT"/>
             </constraints>
-            <simulatedNavigationBarMetrics key="simulatedTopBarMetrics" prompted="NO"/>
         </view>
     </objects>
 </document>
