+++ AppDevPods/AppDevUIKit/ADKGradientEnum.h
@@ -1,0 +1,19 @@
+//
+//  ADKGradientEnum.h
+//  AppDevKit
+//
+//  Created by  Chih Feng Sung on 1/30/19.
+//  Copyright © 2019 Yahoo. All rights reserved.
+//
+
+#ifndef ADKGradientEnum_h
+#define ADKGradientEnum_h
+
+typedef NS_ENUM(NSUInteger, ADKBlendsType) {
+    ADKBlendsTypeFromTopToBottom = 0,
+    ADKBlendsTypeFromLeftToRight,
+    ADKBlendsTypeFromLeftTopToRightBottom,
+    ADKBlendsTypeFromRightTopToLeftBottom
+};
+
+#endif /* ADKGradientEnum_h */
