+++ AppDevPods/AppDevUIKit/ADKGradientView.h
@@ -14,13 +14,14 @@
  */
 
 #import <UIKit/UIKit.h>
+#import "ADKGradientEnum.h"
 
-typedef NS_ENUM(NSUInteger, ADKBlendsType) {
-    ADKBlendsTypeFromTopToBottom = 0,
-    ADKBlendsTypeFromLeftToRight,
-    ADKBlendsTypeFromLeftTopToRightBottom,
-    ADKBlendsTypeFromRightTopToLeftBottom
-};
+//typedef NS_ENUM(NSUInteger, ADKBlendsType) {
+//    ADKBlendsTypeFromTopToBottom = 0,
+//    ADKBlendsTypeFromLeftToRight,
+//    ADKBlendsTypeFromLeftTopToRightBottom,
+//    ADKBlendsTypeFromRightTopToLeftBottom
+//};
 
 @interface ADKGradientView : UIView
 
@@ -41,6 +42,12 @@ typedef NS_ENUM(NSUInteger, ADKBlendsType) {
  *
  */
 @property (nonatomic, assign) ADKBlendsType blendsType;
+
+/**
+ * @brief Asking current view needs to redraw itself. if you change any setting after initialized, you might need to redraw view to display the newest setting.
+ *
+ */
+- (void)redrawView;
 
 @end
